Resolved a bug surrounding PHPs handling of unicode characters + also added the ancillary HTML code required to implement encrypting
Encrypting functionalities as shown in the HTML is currently disabled whilst the PHP element of this is built. This functionality is likely to not be finished for a while, as the deciphering element was purely designed to be utilised alongside the ns-pCB Web Decrypter

New function "unichr" assists with the way the decryption process works.